requirements.txt
numpy
matplotlib
tensorflow-gpu
h5py
tensorflow-datasets==1.2.0